feat(3d): promote connectivity-ordered engine to public API, own module (issue #256/#255 Phase 3)
Phase 1 (PR #386), Phase 2 (PR #387), and the new-island ring-entry-
direction fix (PR #388) together cleared every Phase 3 go/no-go
criterion recorded in ROADMAP.md: every RFC known-broken topology
recovered, a clear improvement on issue #277's real-molecule
population with zero new breakage anywhere measured, gross clash and
stereo-violation counts hold, wall time is comparable, and quality is
atom-order-stable.

Moves generate_coords_connectivity_ordered and its ~13 internal
helpers from dg.rs into a new dg_connectivity_ordered.rs module (a
pure mechanical split, noted as a prerequisite before Phase 3 in
ROADMAP.md -- dg.rs had grown to ~2350 lines mixing two engines).
Promotes the entry point from pub(crate) to pub, and the five
low-level helpers it shares with dg.rs's own legacy generate_coords
(ideal_bond_len, ideal_angle, connected_components, perpendicular_to,
rotate_around_axis) from private to pub(crate).

This is Phase 3 option (c) from the RFC: ship the new engine as an
available, independently-selectable public API, not a default-behavior
switch (option a) or topology-based routing (option b). generate_coords
itself, and every production caller that uses it (generate_coords_etkdg,
embed_pipeline_v2, ...), is unchanged -- no existing behavior changes
for any current caller. The bigger, less-reversible routing decisions
remain separate and not made here.

Verified behaviorally inert: the Phase 2 differential-evaluation
harness produces byte-identical aggregate numbers before and after the
split. 46/46 dg::/dg_connectivity_ordered:: tests pass (same total as
before the split), workspace clippy/fmt clean.

Issues #255/#256/#277 stay open per this project's standing convention
-- this makes the engine available, it does not close out the issues'
own broader scope.
